Putting in new unified IntPoint system

+++ Slicer/SlicerLayer.cs
@@ -19,7 +19,7 @@ You should have received a copy of the GNU Affero General Public License
 along with this program.  If not, see <http://www.gnu.org/licenses/>.
 */
 
-using ClipperLib;
+using MSClipperLib;
 using System.Collections.Generic;
 
 namespace MatterHackers.MatterSlice
@@ -54,9 +54,9 @@ namespace MatterHackers.MatterSlice
 		public List<SlicePerimeterSegment> SegmentList = new List<SlicePerimeterSegment>();
 
 		private Polygons openPolygonList = new Polygons();
-		private int z;
+		private long z;
 
-		public MeshProcessingLayer(int z)
+		public MeshProcessingLayer(long z)
 		{
 			this.z = z;
 		}
@@ -67,7 +67,7 @@ namespace MatterHackers.MatterSlice
 			SegmentList = CreateSegmentListFromString(segmentListData);
 		}
 
-		public int Z { get { return z; } }
+		public long Z { get { return z; } }
 
 		public static List<SlicePerimeterSegment> CreateSegmentListFromString(string segmentListData)
 		{
